+++ src/main/java/br/com/maurigvs/surveyapi/repository/ChoiceRepository.java
@@ -1,10 +1,9 @@
 package br.com.maurigvs.surveyapi.repository;
 
-import org.springframework.data.jpa.repository.JpaRepository;
-
 import br.com.maurigvs.surveyapi.model.Choice;
+import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface ChoiceRepository extends JpaRepository<Choice, Integer> {
+public interface ChoiceRepository extends JpaRepository<Choice, Long> {
 }
